refactored code, optimised saving the location data,

+++ iaeste_table/views.py
@@ -1,28 +1,26 @@
-import folium
 from django.shortcuts import render
-from scripts.mapHandler import *
+from scripts.map_handler import *
 from datetime import date
 import os
 
 
-def mainList(request):
-    from scripts import loadDataFromIAESTESpreadsheet as load
+def main_list(request):
+    from scripts import load_data_from_iaeste_spreadsheet as load
     load.DataLoader()
     from iaeste_table.models import Offer
     offer_list = Offer.objects.all()
 
-    return render(request, "mainList_view_1.html", {"offer_list": offer_list, "offer_list_len": len(offer_list)})
+    return render(request, "offers_list.html", {"offer_list": offer_list, "offer_list_len": len(offer_list)})
 
 
-def detail(request, RefNo):
+def detail(request, ref_no):
     from iaeste_table.models import Offer
-    offer = Offer.objects.get(RefNo=RefNo)
-    return render(request, "detail_view_1.html", {"offer": offer})
+    offer = Offer.objects.get(ref_no=ref_no)
+    return render(request, "offer_detail.html", {"offer": offer})
 
 
-def saveMapToFile(directory, file_to_save, folium_map_as_html):
-    ## create a directory if it does not exist
-    # Check if the directory exists
+def save_map_to_file(directory, file_to_save, folium_map_as_html):
+    # create a directory if it does not exist
     if not os.path.exists(directory):
         # If it doesn't exist, create it
         os.makedirs(directory)
@@ -31,12 +29,12 @@ def saveMapToFile(directory, file_to_save, folium_map_as_html):
 
 
 def get_map(request):
-    currentDate = date.today()
+    current_date = date.today()
     # currentDate = date(2023, 12, 10)  # debugging
-    save_directory = 'savedMaps/'
+    save_directory = 'saved_maps/'
     filename = "map_on_"
     save_format = '.txt'
-    save_date = str(currentDate)
+    save_date = str(current_date)
 
     file_to_save = save_directory + filename + save_date + save_format
 
@@ -45,21 +43,23 @@ def get_map(request):
         with open(file_to_save, 'r', encoding='utf-8') as file:
             folium_map_as_html = file.read().replace('\n', '')
 
-        return render(request, "map_view_1.html", {"mapOfOffers": folium_map_as_html})
+        return render(request, "offers_map.html", {"mapOfOffers": folium_map_as_html})
     else:
-        # delete the existing file
-        for f in os.listdir(save_directory):
-            os.remove(os.path.join(save_directory, f))
+        directory_exists = os.path.isdir(save_directory)
+        if directory_exists:
+            for f in os.listdir(save_directory):
+                os.remove(os.path.join(save_directory, f))  # delete the existing file
 
-        currentUrl = request.build_absolute_uri()
-        folium_map_as_html = createMapForOffers(currentUrl)
-        saveMapToFile(save_directory, file_to_save, folium_map_as_html)
-        return render(request, "map_view_1.html", {"mapOfOffers": folium_map_as_html})
+        current_url = request.build_absolute_uri()
+        folium_map_as_html = create_map_for_offers(current_url)
+        save_map_to_file(save_directory, file_to_save, folium_map_as_html)
 
-
-def aboutProject(request):
-    return render(request, "aboutProject_view_1.html")
+        return render(request, "offers_map.html", {"mapOfOffers": folium_map_as_html})
 
 
-def aboutAuthor(request):
-    return render(request, "aboutAuthor_view_1.html")
+def about_project(request):
+    return render(request, "about_project.html")
+
+
+def about_author(request):
+    return render(request, "about_author.html")
